223 Moves timeline util functions to timelineUtils file
